requirements.txt
s3fs~=2024.3.1
pandas~=2.2.2
requests~=2.31.0
snowflake-connector-python==3.8.1
